224: Cleanup. Rename UserStatus to UserStatusEnum
