+++ api/src/main/java/io/oxalate/backend/api/request/UserStatusRequest.java
@@ -1,7 +1,7 @@
 package io.oxalate.backend.api.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.oxalate.backend.api.UserStatus;
+import io.oxalate.backend.api.UserStatusEnum;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
@@ -12,5 +12,5 @@ public class UserStatusRequest {
     @NotBlank
     @Schema(description = "User status to be set", example = "LOCKED", requiredMode = Schema.RequiredMode.REQUIRED)
     @JsonProperty("status")
-    private UserStatus status;
+    private UserStatusEnum status;
 }
