+++ service/src/main/java/io/oxalate/backend/controller/AuthController.java
@@ -5,9 +5,9 @@ import static io.oxalate.backend.api.AuditLevel.INFO;
 import static io.oxalate.backend.api.AuditLevel.WARN;
 import static io.oxalate.backend.api.SecurityConstants.JWT_TOKEN;
 import io.oxalate.backend.api.UpdateStatusEnum;
-import static io.oxalate.backend.api.UserStatus.ACTIVE;
-import static io.oxalate.backend.api.UserStatus.LOCKED;
-import static io.oxalate.backend.api.UserStatus.REGISTERED;
+import static io.oxalate.backend.api.UserStatusEnum.ACTIVE;
+import static io.oxalate.backend.api.UserStatusEnum.LOCKED;
+import static io.oxalate.backend.api.UserStatusEnum.REGISTERED;
 import io.oxalate.backend.api.request.EmailRequest;
 import io.oxalate.backend.api.request.LoginRequest;
 import io.oxalate.backend.api.request.SignupRequest;
