+++ service/src/main/java/io/oxalate/backend/model/User.java
@@ -2,7 +2,7 @@ package io.oxalate.backend.model;
 
 import io.oxalate.backend.api.MembershipStatusEnum;
 import io.oxalate.backend.api.RoleEnum;
-import io.oxalate.backend.api.UserStatus;
+import io.oxalate.backend.api.UserStatusEnum;
 import io.oxalate.backend.api.request.SignupRequest;
 import io.oxalate.backend.api.response.AdminUserResponse;
 import io.oxalate.backend.api.response.ListUserResponse;
@@ -75,7 +75,7 @@ public class User {
     @NotNull
     @Enumerated(EnumType.STRING)
     @Column(name = "status")
-    private UserStatus status;
+    private UserStatusEnum status;
 
     @NotNull
     @Column(name = "phone_number")
@@ -134,7 +134,7 @@ public class User {
         this.roles.add(new Role(RoleEnum.ROLE_USER));
         this.organizedEvents = new ArrayList<>();
         this.participatedEvents = new ArrayList<>();
-        this.status = UserStatus.REGISTERED;
+        this.status = UserStatusEnum.REGISTERED;
         this.registered = Instant.now();
         this.approvedTerms = signupRequest.isApprovedTerms();
         this.language = signupRequest.getLanguage();
