+++ service/src/test/java/io/oxalate/backend/service/EventServiceITC.java
@@ -13,8 +13,8 @@ import static io.oxalate.backend.api.PortalConfigEnum.PaymentConfigEnum.SINGLE_P
 import io.oxalate.backend.api.RoleEnum;
 import static io.oxalate.backend.api.RoleEnum.ROLE_ORGANIZER;
 import static io.oxalate.backend.api.RoleEnum.ROLE_USER;
-import io.oxalate.backend.api.UserStatus;
-import static io.oxalate.backend.api.UserStatus.ACTIVE;
+import io.oxalate.backend.api.UserStatusEnum;
+import static io.oxalate.backend.api.UserStatusEnum.ACTIVE;
 import io.oxalate.backend.api.request.EventRequest;
 import io.oxalate.backend.api.request.PaymentRequest;
 import io.oxalate.backend.model.Event;
@@ -294,7 +294,7 @@ class EventServiceITC extends AbstractIntegrationTest {
         return eventRepository.save(event);
     }
 
-    private User generateUser(UserStatus userStatus, RoleEnum roleEnum) {
+    private User generateUser(UserStatusEnum userStatusEnum, RoleEnum roleEnum) {
         var randomUsername = "test-" + Instant.now()
                                               .toEpochMilli() + "@test.tld";
         var user = User.builder()
@@ -302,7 +302,7 @@ class EventServiceITC extends AbstractIntegrationTest {
                        .password("password")
                        .firstName("Max")
                        .lastName("Mustermann")
-                       .status(userStatus)
+                       .status(userStatusEnum)
                        .phoneNumber("123456789")
                        .privacy(false)
                        .nextOfKin("Maxine Mustermann")
