+++ service/src/test/java/io/oxalate/backend/service/PaymentServiceITC.java
@@ -6,8 +6,8 @@ import static io.oxalate.backend.api.PortalConfigEnum.PAYMENT;
 import static io.oxalate.backend.api.PortalConfigEnum.PaymentConfigEnum.SINGLE_PAYMENT_ENABLED;
 import io.oxalate.backend.api.RoleEnum;
 import static io.oxalate.backend.api.RoleEnum.ROLE_USER;
-import io.oxalate.backend.api.UserStatus;
-import static io.oxalate.backend.api.UserStatus.ACTIVE;
+import io.oxalate.backend.api.UserStatusEnum;
+import static io.oxalate.backend.api.UserStatusEnum.ACTIVE;
 import io.oxalate.backend.api.request.PaymentRequest;
 import io.oxalate.backend.model.User;
 import io.oxalate.backend.repository.PaymentRepository;
@@ -77,7 +77,7 @@ class PaymentServiceITC extends AbstractIntegrationTest {
         assertTrue(retrievedPayment.isPresent());
     }
 
-    private User generateUser(UserStatus userStatus, RoleEnum roleEnum) {
+    private User generateUser(UserStatusEnum userStatusEnum, RoleEnum roleEnum) {
         var randomUsername = "test-" + Instant.now()
                                               .toEpochMilli() + "@test.tld";
         var user = User.builder()
@@ -85,7 +85,7 @@ class PaymentServiceITC extends AbstractIntegrationTest {
                        .password("password")
                        .firstName("Max")
                        .lastName("Mustermann")
-                       .status(userStatus)
+                       .status(userStatusEnum)
                        .phoneNumber("123456789")
                        .privacy(false)
                        .nextOfKin("Maxine Mustermann")
